pyproject.toml
[project]
name = "albertschool-llmops"
version = "0.1.0"
description = "Course materails for the LLMOps class for Albert School!"
authors = ["Mathieu Soul <mathieus@sicara.com>"]
readme = "README.md"
requires-python = "==3.11.6"
dependencies = [
    "google-cloud-aiplatform[prediction]>=1.38.0",
    "google-cloud-storage>=2.10.0",
    "pandas>=2.3.2",
    "kfp>=2.0.0",
    "datasets>=4.1.1",
    "ipykernel>=6.30.1",
    "ragas>=0.3.5",
    "rouge-score>=0.1.2",
    "langchain-core>=0.3.76",
    "langchain-google-vertexai>=2.1.2",
    "langchain>=0.3.27",
    "gcsfs>=2025.9.0",
    "tensorboard>=2.20.0",
    "rapidfuzz>=3.14.1",
    "sacrebleu>=2.5.1",
    "langchain-google-genai>=2.1.12",
    "langchain-huggingface>=0.3.1",
    "chainlit>=1.3.1",
    "transformers==4.46.*",
    "accelerate>=1.10.1",
    "trl>=0.17.0",
    "peft==0.13.2",
    "pydantic==2.10.1",
    "typer>=0.19.2",
    "langfuse>=3.8.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "mypy<2.0,>=1.2",
    "pre-commit<3,>=2",
    "pytest<8,>=7",
    "pytest-cov<4,>=3",
    "ruff<1.0.0,>=0.8.0",
]

# Packages configs

## coverage

[tool.coverage.run]
branch = true

[tool.coverage.report]
skip_empty = true
omit = [
    "src/constants.py",
]
exclude_also = [
    # Such code branches cannot be easily tested; that's OK since they should only contain 1 function call
    "if __name__ == \"__main__\":",
]
fail_under = 70.00
precision = 2

## ruff
# Recommended ruff config for now, to be updated as we go along.
[tool.ruff]
target-version = 'py311'

[tool.ruff.lint]
# See all rules at https://docs.astral.sh/ruff/rules/
select = [
    "E",      # pycodestyle
    "W",      # pycodestyle
    "F",      # Pyflakes
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "D",      # pydocstyle
    "I",      # isort
    "SIM",    # flake8-simplify
    "TC",     # flake8-type-checking
    "TID",    # flake8-tidy-imports
    "UP",     # pyupgrade
    "PT",     # flake8-pytest-style
    "RUF",    # Ruff-specific rules
    "FBT001", # flake8-boolean-trap
]

ignore = [
    "E501", # "Line too long"
    # -> line length already regulated by the formatter
    "PT011", # "pytest.raises() should specify expected exception"
    # -> would imply to update tests every time you update exception message
    "SIM102", # "Use a single `if` statement instead of nested `if` statements"
    # -> too restrictive
]

[tool.ruff.lint.pydocstyle]
# Automatically disable rules that are incompatible with Google docstring convention
convention = "google"

[tool.ruff.lint.pycodestyle]
max-doc-length = 88

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-type-checking]
strict = true
runtime-evaluated-base-classes = ["pydantic.BaseModel"]
# Pydantic needs to be able to evaluate types at runtime
# see https://pypi.org/project/flake8-type-checking/ for flake8-type-checking documentation
# see https://beta.ruff.rs/docs/settings/#flake8-type-checking-runtime-evaluated-base-classes for ruff documentation

[tool.ruff.lint.per-file-ignores]
# Allow missing docstrings for tests
"tests/**/*.py" = ["D1"]

## mypy

[tool.mypy]
python_version = "3.11"
# Enable all optional error checking flags, providing stricter type checking; see https://mypy.readthedocs.io/en/stable/getting_started.html#strict-mode-and-configuration
strict = true

# Type-check the interiors of functions without type annotations; if missing, mypy won't check function bodies without type hints, for instance those coming from third-party libraries
check_untyped_defs = true

# Make __init__.py file optional for package definitions; if missing, mypy requires __init__.py at packages roots, see https://mypy.readthedocs.io/en/stable/running_mypy.html#mapping-file-paths-to-modules
explicit_package_bases = true


[[tool.mypy.overrides]]
module = [
# Ignore missing library stubs for the following packages;
# see https://mypy.readthedocs.io/en/stable/running_mypy.html#missing-library-stubs-or-py-typed-marker
# and https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml-file
]
ignore_missing_imports = true
